fix ctcf index

+++ src/epicast/utils/math_utils.py
@@ -11,8 +11,8 @@ def sigmoid(x):
     return 1 / (1 + np.exp(-x))
 
 
-def logit(x, eps=0):
-    x = np.asarray(x, dtype=np.float64)
+def logit(x, eps=0.0):
+    x = np.asarray(x)
     x = np.clip(x, eps, 1 - eps)
     return np.log(x/(1-x))
